Reduce namespace pollution (Windows)
Include directives of Windows-specific headers in header files of
this library caused leakage of Windows-specific declarations and
definitions into the namespaces of unrelated parts of this library
and into the namespaces of the consumers of this library. This
was fixed by moving these Windows-specific include directives
from the library header files into the library source files.
Consequently, Windows-specific implementation code had to
be moved as well from the library header files into the library
source files. This was done by applying the PIMPL (Pointer to
IMPLementation) pattern.